+++ text/src/typography.rs
@@ -6,7 +6,7 @@ use smallvec::{smallvec, SmallVec};
 use unicode_script::{Script, UnicodeScript};
 use unicode_segmentation::UnicodeSegmentation;
 
-use crate::{shaper::ShapeResult, Em, FontSize, Glyph, Overflow, Pixel, TextAlign};
+use crate::{shaper::ShapeResult, Glyph, GlyphUnit, Overflow, TextAlign};
 
 #[derive(Clone, Copy, PartialEq, Eq, Hash)]
 pub enum PlaceLineDirection {
@@ -20,44 +20,32 @@ pub enum PlaceLineDirection {
   BottomToTop,
 }
 
-#[derive(Clone)]
-pub struct TypographyCfg {
-  pub line_height: Option<Em>,
-  pub letter_space: Option<Pixel>,
-  pub text_align: TextAlign,
-  // The size glyphs can place, and hint `TypographyMan` where to early return.
-  // the result of typography may over bounds.
-  pub bounds: Size<Em>,
-  pub line_dir: PlaceLineDirection,
-  pub overflow: Overflow,
-}
-
 /// Trait control how to place glyph inline.
 pub trait InlineCursor {
   /// advance the cursor by a glyph, the `glyph` position is relative to self
   /// before call this method,  and relative to the cursor coordinate after
   /// call.
-  fn advance_glyph(&mut self, glyph: &mut Glyph<Em>, line_offset: Em, origin_text: &str);
+  fn advance_glyph(&mut self, glyph: &mut Glyph, line_offset: GlyphUnit, origin_text: &str);
 
   /// advance the cursor position by a offset em
-  fn advance(&mut self, offset: Em);
+  fn advance(&mut self, offset: GlyphUnit);
 
   /// cursor position relative of inline.
-  fn position(&self) -> Em;
+  fn position(&self) -> GlyphUnit;
 
   fn reset(&mut self);
 
-  fn measure(&self, glyph: &Glyph<Em>, origin_text: &str) -> Em;
+  fn measure(&self, glyph: &Glyph, origin_text: &str) -> GlyphUnit;
 }
 
 #[derive(Default)]
 pub struct VisualLine {
-  pub x: Em,
-  pub y: Em,
-  pub height: Em,
-  pub width: Em,
+  pub x: GlyphUnit,
+  pub y: GlyphUnit,
+  pub height: GlyphUnit,
+  pub width: GlyphUnit,
   /// The glyph position is relative the line x/y
-  pub glyphs: Vec<Glyph<Em>>,
+  pub glyphs: Vec<Glyph>,
 }
 
 pub struct VisualInfos {
@@ -66,18 +54,22 @@ pub struct VisualInfos {
   /// if the typography result over the bounds provide by caller.
   pub over_bounds: bool,
   pub line_dir: PlaceLineDirection,
-  pub visual_width: Em,
-  pub visual_height: Em,
+  pub visual_width: GlyphUnit,
+  pub visual_height: GlyphUnit,
 }
 
 /// Typography the glyphs in a bounds.
 pub struct TypographyMan<Paras> {
-  cfg: TypographyCfg,
+  line_dir: PlaceLineDirection,
+  text_align: TextAlign,
+  line_height: GlyphUnit,
+  bounds: Size<GlyphUnit>,
+  overflow: Overflow,
   /// Not directly use text as inputs, but accept glyphs after text shape
   /// because both simple text and rich text can custom compose its glyph runs
   /// by text reorder result and its style .
   inputs: Paras,
-  inline_cursor: Em,
+  inline_cursor: GlyphUnit,
   visual_lines: SmallVec<[VisualLine; 1]>,
   over_bounds: bool,
 }
@@ -86,8 +78,21 @@ impl<Paras> TypographyMan<Paras>
 where
   Paras: DoubleEndedIterator<Item = SmallVec<[InputRun; 1]>>,
 {
-  pub fn new(inputs: Paras, cfg: TypographyCfg) -> Self {
-    Self { cfg, inputs, inline_cursor: Em::ZERO, visual_lines: smallvec![], over_bounds: false }
+  pub fn new(
+    inputs: Paras, line_dir: PlaceLineDirection, text_align: TextAlign, line_height: GlyphUnit,
+    bounds: Size<GlyphUnit>, overflow: Overflow,
+  ) -> Self {
+    Self {
+      line_dir,
+      text_align,
+      line_height,
+      bounds,
+      overflow,
+      inputs,
+      inline_cursor: GlyphUnit::ZERO,
+      visual_lines: smallvec![],
+      over_bounds: false,
+    }
   }
 
   pub fn typography_all(mut self) -> VisualInfos {
@@ -95,7 +100,7 @@ where
       self.consume_paragraph(p);
     }
 
-    if self.cfg.line_dir.is_reverse() {
+    if self.line_dir.is_reverse() {
       self.visual_lines.reverse();
     }
 
@@ -105,26 +110,26 @@ where
     VisualInfos {
       visual_width,
       visual_height,
-      text_align: self.cfg.text_align,
+      text_align: self.text_align,
       visual_lines: self.visual_lines,
       over_bounds: self.over_bounds,
-      line_dir: self.cfg.line_dir,
+      line_dir: self.line_dir,
     }
   }
 
-  fn adjust_lines(&mut self, visual_width: Em, visual_height: Em) {
-    let text_align = self.cfg.text_align;
+  fn adjust_lines(&mut self, visual_width: GlyphUnit, visual_height: GlyphUnit) {
+    let text_align = self.text_align;
 
-    match self.cfg.line_dir {
+    match self.line_dir {
       PlaceLineDirection::LeftToRight | PlaceLineDirection::RightToLeft => {
-        let mut x_offset = Em::absolute(0.);
+        let mut x_offset = GlyphUnit::ZERO;
         self.visual_lines.iter_mut().for_each(move |l| {
           l.x = x_offset;
           x_offset += l.width;
         });
       }
       PlaceLineDirection::TopToBottom | PlaceLineDirection::BottomToTop => {
-        let mut y_offset = Em::absolute(0.);
+        let mut y_offset = GlyphUnit::ZERO;
         self.visual_lines.iter_mut().for_each(move |l| {
           l.y = y_offset;
           y_offset += l.height;
@@ -133,23 +138,18 @@ where
     };
 
     self.visual_lines.iter_mut().for_each(|l| {
-      let (x, y) = text_align_offset(
-        self.cfg.line_dir.is_horizontal(),
-        text_align,
-        visual_width,
-        visual_height,
-        l.width,
-        l.height,
-      );
-      l.x += x;
-      l.y += y;
+      if self.line_dir.is_horizontal() {
+        l.y += text_align_offset(l.height, visual_height, text_align);
+      } else {
+        l.x += text_align_offset(l.width, visual_width, text_align);
+      }
     });
   }
 
-  fn visual_size(&self) -> (Em, Em) {
-    let mut width = Em::ZERO;
-    let mut height = Em::ZERO;
-    if self.cfg.line_dir.is_horizontal() {
+  fn visual_size(&self) -> (GlyphUnit, GlyphUnit) {
+    let mut width = GlyphUnit::ZERO;
+    let mut height = GlyphUnit::ZERO;
+    if self.line_dir.is_horizontal() {
       self.visual_lines.iter().for_each(|l| {
         width += l.width;
         height = height.max(l.height);
@@ -168,16 +168,16 @@ where
   fn consume_paragraph(&mut self, runs: SmallVec<[InputRun; 1]>) -> bool {
     self.begin_line();
 
-    if self.cfg.line_dir.is_horizontal() {
+    if self.line_dir.is_horizontal() {
       let mut cursor = VInlineCursor { pos: self.inline_cursor };
       runs
         .iter()
-        .for_each(|r| self.consume_run_with_letter_space_cursor(&r, &mut cursor));
+        .for_each(|r| self.consume_run_with_letter_space_cursor(r, &mut cursor));
     } else {
       let mut cursor = HInlineCursor { pos: self.inline_cursor };
       runs
         .iter()
-        .for_each(|r| self.consume_run_with_letter_space_cursor(&r, &mut cursor));
+        .for_each(|r| self.consume_run_with_letter_space_cursor(r, &mut cursor));
     }
     self.end_line();
 
@@ -187,66 +187,32 @@ where
   fn consume_run_with_letter_space_cursor(
     &mut self, run: &InputRun, inner_cursor: &mut impl InlineCursor,
   ) {
-    let letter_space = run
-      .letter_space
-      .or(self.cfg.letter_space)
-      .unwrap_or(Pixel::ZERO);
-    if letter_space != Em::ZERO {
-      let mut cursor = LetterSpaceCursor::new(inner_cursor, letter_space.into());
+    if run.letter_space != GlyphUnit::ZERO {
+      let mut cursor = LetterSpaceCursor::new(inner_cursor, run.letter_space);
       self.consume_run(run, &mut cursor);
     } else {
       self.consume_run(run, inner_cursor);
     }
   }
 
-  fn split_word<'a>(
-    &self, run: &'a InputRun,
-  ) -> impl Iterator<Item = impl Iterator<Item = &'a Glyph<Em>> + 'a> + 'a {
-    let text = run.text();
-    let mut reorder_text = String::new();
-
-    // text and glyphs in run may in different order, so we recollect the chars.
-    // and reorder_text may smaller then src text when have composited glyph,
-    // like '👨‍👩‍👦‍👦'
-    reorder_text.reserve(text.len());
-    run.glyphs().iter().for_each(|gh| {
-      reorder_text.push(
-        text[gh.cluster as usize..]
-          .chars()
-          .next()
-          .unwrap(),
-      )
-    });
-
-    let mut it = reorder_text.split_word_bounds();
-    let mut base = 0;
-    let mut words = vec![];
-    for text in it.by_ref() {
-      let char_cnt = text.chars().count();
-      words.push(base..char_cnt + base);
-      base += char_cnt;
-    }
-    words.into_iter().map(move |rg| {
-      rg.into_iter()
-        .map(move |idx| run.glyphs().get(idx).unwrap())
-    })
-  }
-
   fn consume_run(&mut self, run: &InputRun, cursor: &mut impl InlineCursor) {
-    let font_size = run.font_size.into_em();
+    let font_size = run.font_size_factor * GlyphUnit::PIXELS_PER_EM as f32;
     let text = run.text();
     let base = run.range.start as u32;
-    let line_offset = (font_size - Em::absolute(1.)) / 2.;
-    let is_auto_wrap = self.cfg.overflow.is_auto_wrap();
+    let em = GlyphUnit::from_pixel(font_size);
+    let line_offset = (em - GlyphUnit::STANDARD_EM) / 2.;
+    let is_auto_wrap = self.overflow.is_auto_wrap();
 
     let verify_line_height = |this: &mut Self| {
       let line = this.visual_lines.last_mut().unwrap();
-      if this.cfg.line_dir.is_horizontal() {
-        line.width = line.width.max(font_size)
+      if this.line_dir.is_horizontal() {
+        line.width = line.width.max(em)
       } else {
-        line.height = line.height.max(font_size)
+        line.height = line.height.max(em)
       }
     };
+    (verify_line_height)(self);
+
     let new_line = |this: &mut Self, cursor: &mut dyn InlineCursor| {
       this.end_line();
       this.begin_line();
@@ -254,34 +220,19 @@ where
       cursor.reset();
     };
 
-    let words = self
-      .split_word(run)
-      .map(|it| {
-        it.cloned().map(|mut g| {
-          g.scale(font_size.value());
-          g
-        })
-      })
-      .map(|it| {
-        let word = it.collect::<Vec<_>>();
-        let width = word
-          .iter()
-          .fold(Em::ZERO, |acc, glyph| acc + cursor.measure(glyph, text));
-        (width, word)
-      })
-      .collect::<Vec<_>>();
+    for word in run.word_glyphs() {
+      let width: GlyphUnit = word
+        .clone()
+        .fold(GlyphUnit::ZERO, |acc, g| acc + cursor.measure(&g, text));
 
-    (verify_line_height)(self);
-    for (width, word) in words {
       if is_auto_wrap
-        && self.inline_cursor != Em::ZERO
+        && self.inline_cursor != GlyphUnit::ZERO
         && self.is_over_line_bound(width + self.inline_cursor)
       {
         new_line(self, cursor);
       }
 
-      let mut word = word.iter().peekable();
-
+      let mut word = word.peekable();
       while let Some(g) = word.peek() {
         let mut at = (*g).clone();
 
@@ -289,7 +240,7 @@ where
 
         at.cluster += base;
 
-        if self.inline_cursor == Em::ZERO
+        if self.inline_cursor == GlyphUnit::ZERO
           || !is_auto_wrap
           || !self.is_over_line_bound(cursor.position())
         {
@@ -303,7 +254,7 @@ where
     }
   }
 
-  fn push_glyph(&mut self, g: Glyph<Em>) {
+  fn push_glyph(&mut self, g: Glyph) {
     let line = self.visual_lines.last_mut();
     line.unwrap().glyphs.push(g)
   }
@@ -313,113 +264,110 @@ where
   fn end_line(&mut self) {
     let line = self.visual_lines.last_mut().unwrap();
     // we will reorder the line after consumed all inputs.
-    if self.cfg.line_dir.is_horizontal() {
+    if self.line_dir.is_horizontal() {
       line.height = self.inline_cursor;
-      if let Some(line_height) = self.cfg.line_height {
-        line.width = line_height;
-      }
+      line.width = self.line_height;
     } else {
       line.width = self.inline_cursor;
-      if let Some(line_height) = self.cfg.line_height {
-        line.height = line_height;
-      }
+      line.height = self.line_height;
     }
 
     self.over_bounds |= self.is_over_line_bound(self.inline_cursor);
     self.over_bounds |= self.is_last_line_over();
-    self.inline_cursor = Em::ZERO;
+    self.inline_cursor = GlyphUnit::ZERO;
   }
 
-  fn is_over_line_bound(&self, position: Em) -> bool {
-    if self.cfg.text_align == TextAlign::Center {
+  fn is_over_line_bound(&self, position: GlyphUnit) -> bool {
+    if self.text_align == TextAlign::Center {
       return false;
     }
 
-    let eps: Em = Em(0.00001_f32);
-    if self.cfg.line_dir.is_horizontal() {
-      self.cfg.bounds.height + eps <= position
+    if self.line_dir.is_horizontal() {
+      self.bounds.height <= position
     } else {
-      self.cfg.bounds.width + eps <= position
+      self.bounds.width <= position
     }
   }
 
   fn is_last_line_over(&self) -> bool {
-    if self.cfg.line_dir.is_horizontal() {
-      self.cfg.bounds.width
+    if self.line_dir.is_horizontal() {
+      self.bounds.width
         < self
           .visual_lines
           .iter()
-          .fold(Em::ZERO, |acc, l| acc + l.width)
+          .fold(GlyphUnit::ZERO, |acc, l| acc + l.width)
     } else {
-      self.cfg.bounds.height
+      self.bounds.height
         < self
           .visual_lines
           .iter()
-          .fold(Em::ZERO, |acc, l| acc + l.height)
+          .fold(GlyphUnit::ZERO, |acc, l| acc + l.height)
     }
   }
 }
 
 pub struct InputRun {
   pub(crate) shape_result: Sc<ShapeResult>,
-  pub(crate) font_size: FontSize,
-  pub(crate) letter_space: Option<Pixel>,
+  /// The factor relative to the standard size.
+  pub(crate) font_size_factor: f32,
+  pub(crate) letter_space: GlyphUnit,
   pub(crate) range: Range<usize>,
+  reorder_text: String,
 }
 
 pub struct HInlineCursor {
-  pub pos: Em,
+  pub pos: GlyphUnit,
 }
 
 pub struct VInlineCursor {
-  pub pos: Em,
+  pub pos: GlyphUnit,
 }
 
 pub struct LetterSpaceCursor<'a, I> {
   inner_cursor: &'a mut I,
-  letter_space: Em,
+  letter_space: GlyphUnit,
 }
 
 impl<'a, I> LetterSpaceCursor<'a, I> {
-  pub fn new(inner_cursor: &'a mut I, letter_space: Em) -> Self {
+  pub fn new(inner_cursor: &'a mut I, letter_space: GlyphUnit) -> Self {
     Self { inner_cursor, letter_space }
   }
 }
 
 impl InlineCursor for HInlineCursor {
-  fn advance_glyph(&mut self, g: &mut Glyph<Em>, line_offset: Em, _: &str) {
+  fn advance_glyph(&mut self, g: &mut Glyph, line_offset: GlyphUnit, _: &str) {
     g.x_offset += self.pos;
     g.y_offset += line_offset;
     self.pos = g.x_offset + g.x_advance;
   }
 
-  fn measure(&self, glyph: &Glyph<Em>, _origin_text: &str) -> Em { glyph.x_advance }
+  fn measure(&self, glyph: &Glyph, _origin_text: &str) -> GlyphUnit { glyph.x_advance }
 
-  fn advance(&mut self, c: Em) { self.pos += c; }
+  fn advance(&mut self, c: GlyphUnit) { self.pos += c; }
 
-  fn position(&self) -> Em { self.pos }
+  fn position(&self) -> GlyphUnit { self.pos }
 
-  fn reset(&mut self) { self.pos = Em::ZERO; }
+  fn reset(&mut self) { self.pos = GlyphUnit::ZERO; }
 }
 
 impl InlineCursor for VInlineCursor {
-  fn advance_glyph(&mut self, g: &mut Glyph<Em>, line_offset: Em, _: &str) {
+  fn advance_glyph(&mut self, g: &mut Glyph, line_offset: GlyphUnit, _: &str) {
     g.x_offset += line_offset;
     g.y_offset += self.pos;
     self.pos = g.y_offset + g.y_advance;
   }
 
-  fn advance(&mut self, c: Em) { self.pos += c; }
+  fn advance(&mut self, c: GlyphUnit) { self.pos += c; }
 
-  fn measure(&self, glyph: &Glyph<Em>, _origin_text: &str) -> Em { glyph.y_advance }
+  fn measure(&self, glyph: &Glyph, _origin_text: &str) -> GlyphUnit { glyph.y_advance }
 
-  fn position(&self) -> Em { self.pos }
+  fn position(&self) -> GlyphUnit { self.pos }
 
-  fn reset(&mut self) { self.pos = Em::ZERO; }
+  fn reset(&mut self) { self.pos = GlyphUnit::ZERO; }
 }
 
 impl<'a, I: InlineCursor> InlineCursor for LetterSpaceCursor<'a, I> {
-  fn advance_glyph(&mut self, g: &mut Glyph<Em>, line_offset: Em, origin_text: &str) {
+  fn advance_glyph(&mut self, g: &mut Glyph, line_offset: GlyphUnit, origin_text: &str) {
     let cursor = &mut self.inner_cursor;
     cursor.advance_glyph(g, line_offset, origin_text);
 
@@ -432,7 +380,7 @@ impl<'a, I: InlineCursor> InlineCursor for LetterSpaceCursor<'a, I> {
     }
   }
 
-  fn measure(&self, glyph: &Glyph<Em>, origin_text: &str) -> Em {
+  fn measure(&self, glyph: &Glyph, origin_text: &str) -> GlyphUnit {
     let mut advance = self.inner_cursor.measure(glyph, origin_text);
 
     let c = origin_text[glyph.cluster as usize..]
@@ -446,9 +394,9 @@ impl<'a, I: InlineCursor> InlineCursor for LetterSpaceCursor<'a, I> {
     advance
   }
 
-  fn advance(&mut self, c: Em) { self.inner_cursor.advance(c) }
+  fn advance(&mut self, c: GlyphUnit) { self.inner_cursor.advance(c) }
 
-  fn position(&self) -> Em { self.inner_cursor.position() }
+  fn position(&self) -> GlyphUnit { self.inner_cursor.position() }
 
   fn reset(&mut self) { self.inner_cursor.reset(); }
 }
@@ -464,21 +412,18 @@ impl PlaceLineDirection {
 }
 
 impl VisualLine {
-  pub fn line_height(&self, line_dir: PlaceLineDirection) -> Em {
+  pub fn line_height(&self, line_dir: PlaceLineDirection) -> GlyphUnit {
     if line_dir.is_horizontal() { self.width } else { self.height }
   }
 }
 
 pub(crate) fn text_align_offset(
-  is_horizontal: bool, text_align: TextAlign, bound_width: Em, bound_height: Em, visual_width: Em,
-  visual_height: Em,
-) -> (Em, Em) {
-  match (text_align, is_horizontal) {
-    (TextAlign::Start, _) => (Em::ZERO, Em::ZERO),
-    (TextAlign::Center, true) => (Em::ZERO, (bound_height - visual_height) / 2.),
-    (TextAlign::Center, false) => ((bound_width - visual_width) / 2., Em::ZERO),
-    (TextAlign::End, true) => (Em::ZERO, bound_height - visual_height),
-    (TextAlign::End, false) => (bound_width - visual_width, Em::ZERO),
+  content: GlyphUnit, container: GlyphUnit, text_align: TextAlign,
+) -> GlyphUnit {
+  match text_align {
+    TextAlign::Start => GlyphUnit::ZERO,
+    TextAlign::Center => (container - content) / 2.,
+    TextAlign::End => container - content,
   }
 }
 
@@ -508,9 +453,39 @@ fn letter_spacing_char(c: char) -> bool {
 }
 
 impl InputRun {
+  pub(crate) fn new(
+    shape_result: Sc<ShapeResult>, font_size_factor: f32, letter_space: GlyphUnit,
+    range: Range<usize>,
+  ) -> Self {
+    let text: &str = &shape_result.text;
+    // text and glyphs in run may in different order, so we recollect the chars.
+    // and reorder_text may smaller then src text when have composited glyph,
+    // like '👨‍👩‍👦‍👦'
+    let reorder_text = shape_result
+      .glyphs
+      .iter()
+      .filter_map(|gh| text[gh.cluster as usize..].chars().next())
+      .collect();
+    Self { shape_result, font_size_factor, letter_space, range, reorder_text }
+  }
+
   #[inline]
   fn text(&self) -> &str { &self.shape_result.text }
 
-  #[inline]
-  fn glyphs(&self) -> &[Glyph<Em>] { &self.shape_result.glyphs }
+  fn word_glyphs(&self) -> impl Iterator<Item = impl Iterator<Item = Glyph> + Clone + '_> + '_ {
+    let Self { reorder_text, font_size_factor, shape_result, .. } = self;
+    let font_size = *font_size_factor * GlyphUnit::PIXELS_PER_EM as f32;
+    reorder_text
+      .split_word_bounds()
+      .scan(0, move |init, w| {
+        let base = *init;
+        *init += w.chars().count();
+        Some(
+          w.chars()
+            .enumerate()
+            .filter_map(move |(idx, _)| shape_result.glyphs.get(base + idx))
+            .map(move |g| g.clone().cast_to(font_size)),
+        )
+      })
+  }
 }