+++ themes/material/src/lib.rs
@@ -415,9 +415,9 @@ pub fn typography_theme(
   TypographyTheme {
     display_large: TextTheme {
       text: CowArc::owned(TextStyle {
-        line_height: Some(Em::from_pixel(64.0.into())),
-        font_size: FontSize::Pixel(57.0.into()),
-        letter_space: Some(0.0.into()),
+        line_height: 64.,
+        font_size: 57.,
+        letter_space: 0.,
         font_face: regular_face.clone(),
         overflow: Overflow::Clip,
       }),
@@ -425,9 +425,9 @@ pub fn typography_theme(
     },
     display_medium: TextTheme {
       text: CowArc::owned(TextStyle {
-        line_height: Some(Em::from_pixel(52.0.into())),
-        font_size: FontSize::Pixel(45.0.into()),
-        letter_space: Some(0.0.into()),
+        line_height: 52.0,
+        font_size: 45.0,
+        letter_space: 0.,
         font_face: regular_face.clone(),
         overflow: Overflow::Clip,
       }),
@@ -435,9 +435,9 @@ pub fn typography_theme(
     },
     display_small: TextTheme {
       text: CowArc::owned(TextStyle {
-        line_height: Some(Em::from_pixel(44.0.into())),
-        font_size: FontSize::Pixel(36.0.into()),
-        letter_space: Some(0.0.into()),
+        line_height: 44.0,
+        font_size: 36.0,
+        letter_space: 0.,
         font_face: regular_face.clone(),
         overflow: Overflow::Clip,
       }),
@@ -445,9 +445,9 @@ pub fn typography_theme(
     },
     headline_large: TextTheme {
       text: CowArc::owned(TextStyle {
-        line_height: Some(Em::from_pixel(40.0.into())),
-        font_size: FontSize::Pixel(32.0.into()),
-        letter_space: Some(0.0.into()),
+        line_height: 40.0,
+        font_size: 32.0,
+        letter_space: 0.,
         font_face: regular_face.clone(),
         overflow: Overflow::Clip,
       }),
@@ -455,9 +455,9 @@ pub fn typography_theme(
     },
     headline_medium: TextTheme {
       text: CowArc::owned(TextStyle {
-        line_height: Some(Em::from_pixel(36.0.into())),
-        font_size: FontSize::Pixel(28.0.into()),
-        letter_space: Some(0.0.into()),
+        line_height: 36.0,
+        font_size: 28.0,
+        letter_space: 0.,
         font_face: regular_face.clone(),
         overflow: Overflow::Clip,
       }),
@@ -465,9 +465,9 @@ pub fn typography_theme(
     },
     headline_small: TextTheme {
       text: CowArc::owned(TextStyle {
-        line_height: Some(Em::from_pixel(32.0.into())),
-        font_size: FontSize::Pixel(24.0.into()),
-        letter_space: Some(0.0.into()),
+        line_height: 32.0,
+        font_size: 24.0,
+        letter_space: 0.,
         font_face: regular_face.clone(),
         overflow: Overflow::Clip,
       }),
@@ -475,9 +475,9 @@ pub fn typography_theme(
     },
     title_large: TextTheme {
       text: CowArc::owned(TextStyle {
-        line_height: Some(Em::from_pixel(28.0.into())),
-        font_size: FontSize::Pixel(22.0.into()),
-        letter_space: Some(0.0.into()),
+        line_height: 28.0,
+        font_size: 22.0,
+        letter_space: 0.,
         font_face: medium_face.clone(),
         overflow: Overflow::Clip,
       }),
@@ -485,9 +485,9 @@ pub fn typography_theme(
     },
     title_medium: TextTheme {
       text: CowArc::owned(TextStyle {
-        line_height: Some(Em::from_pixel(24.0.into())),
-        font_size: FontSize::Pixel(16.0.into()),
-        letter_space: Some(0.15.into()),
+        line_height: 24.0,
+        font_size: 16.0,
+        letter_space: 0.15,
         font_face: medium_face.clone(),
         overflow: Overflow::Clip,
       }),
@@ -495,9 +495,9 @@ pub fn typography_theme(
     },
     title_small: TextTheme {
       text: CowArc::owned(TextStyle {
-        line_height: Some(Em::from_pixel(20.0.into())),
-        font_size: FontSize::Pixel(14.0.into()),
-        letter_space: Some(0.1.into()),
+        line_height: 20.0,
+        font_size: 14.0,
+        letter_space: 0.1,
         font_face: medium_face.clone(),
         overflow: Overflow::Clip,
       }),
@@ -505,9 +505,9 @@ pub fn typography_theme(
     },
     label_large: TextTheme {
       text: CowArc::owned(TextStyle {
-        line_height: Some(Em::from_pixel(20.0.into())),
-        font_size: FontSize::Pixel(14.0.into()),
-        letter_space: Some(0.1.into()),
+        line_height: 20.0,
+        font_size: 14.0,
+        letter_space: 0.1,
         font_face: medium_face.clone(),
         overflow: Overflow::Clip,
       }),
@@ -515,9 +515,9 @@ pub fn typography_theme(
     },
     label_medium: TextTheme {
       text: CowArc::owned(TextStyle {
-        line_height: Some(Em::from_pixel(16.0.into())),
-        font_size: FontSize::Pixel(12.0.into()),
-        letter_space: Some(0.5.into()),
+        line_height: 16.0,
+        font_size: 12.0,
+        letter_space: 0.5,
         font_face: medium_face.clone(),
         overflow: Overflow::Clip,
       }),
@@ -525,9 +525,9 @@ pub fn typography_theme(
     },
     label_small: TextTheme {
       text: CowArc::owned(TextStyle {
-        line_height: Some(Em::from_pixel(16.0.into())),
-        font_size: FontSize::Pixel(11.0.into()),
-        letter_space: Some(0.5.into()),
+        line_height: 16.0,
+        font_size: 11.0,
+        letter_space: 0.5,
         font_face: medium_face,
         overflow: Overflow::Clip,
       }),
@@ -535,9 +535,9 @@ pub fn typography_theme(
     },
     body_large: TextTheme {
       text: CowArc::owned(TextStyle {
-        line_height: Some(Em::from_pixel(24.0.into())),
-        font_size: FontSize::Pixel(16.0.into()),
-        letter_space: Some(0.5.into()),
+        line_height: 24.0,
+        font_size: 16.0,
+        letter_space: 0.5,
         font_face: regular_face.clone(),
         overflow: Overflow::Clip,
       }),
@@ -545,9 +545,9 @@ pub fn typography_theme(
     },
     body_medium: TextTheme {
       text: CowArc::owned(TextStyle {
-        line_height: Some(Em::from_pixel(20.0.into())),
-        font_size: FontSize::Pixel(14.0.into()),
-        letter_space: Some(0.25.into()),
+        line_height: 20.0,
+        font_size: 14.0,
+        letter_space: 0.25,
         font_face: regular_face.clone(),
         overflow: Overflow::Clip,
       }),
@@ -555,9 +555,9 @@ pub fn typography_theme(
     },
     body_small: TextTheme {
       text: CowArc::owned(TextStyle {
-        line_height: Some(Em::from_pixel(16.0.into())),
-        font_size: FontSize::Pixel(12.0.into()),
-        letter_space: Some(0.4.into()),
+        line_height: 16.0,
+        font_size: 12.0,
+        letter_space: 0.4,
         font_face: regular_face,
         overflow: Overflow::Clip,
       }),